Removed private helper classes from CalendarWidget
